Implement data sorting for all
